Move machine states in dedicated package
